hf_deploy/requirements_hf.txt
pypdf
faiss-cpu
sentence-transformers
gradio
numpy
